fix: DRAW_LINE segmentation for large diagonals (dx,dy > 127)
- Add VLINE_DX_REMAINING symbol allocation in mod.rs for segment 2 dx storage
- Fix segment 1 clamping: add DLW_SEG1_DY_NO_CLAMP and DLW_SEG1_DX_NO_CLAMP to properly initialize A/B when values are in valid range (-128..127)
- Implement segment 2 dx calculation (mirrors dy calculation logic)
- Fix register order in segment 2: A=dy (not dx), B=dx (not dy) per Draw_Line_d signature
- testdiagonal now renders correctly: diagonal from (-90,-90) to (90,90) draws as 2 diagonal segments (127,127) + (53,53)

Without DLW_SEG1_*_NO_CLAMP: registers were uninitialized for values in range, causing vertical/horizontal lines
Without correct dx calculation: segment 2 was always vertical
Without correct register order: segment 2 was perpendicular to intended direction

+++ core/src/codegen.rs
@@ -1096,7 +1096,10 @@ fn opt_stmt(s: &Stmt) -> Stmt {
 }
 
 const INT_MASK: i32 = 0xFFFF; // unify 16-bit integer model across backends
-fn trunc16(v: i32) -> i32 { v & INT_MASK }
+fn trunc16(v: i32) -> i32 { 
+    // Truncate to 16-bit SIGNED range, preserving negative sign
+    ((v & INT_MASK) as i16) as i32
+}
 
 fn opt_expr(e: &Expr) -> Expr {
     match e {
